install_requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
breadability==0.1.20
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
ffpyplayer==4.5.3
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==3.11.3
keyboard==0.13.5
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
namex==0.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
onnxruntime==1.22.1
openai-whisper==20250625
openwakeword==0.6.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
srt==3.5.3
sumy==0.11.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
